Move stuff around to decouple logic
